+++ data/core/src/commonTest/kotlin/ly/david/musicsearch/data/core/place/CoordinatesTest.kt
@@ -1,7 +1,5 @@
-package ly.david.musicsearch.domain.place
+package ly.david.musicsearch.data.core.place
 
-import ly.david.musicsearch.data.core.place.CoordinatesUiModel
-import ly.david.musicsearch.data.core.place.formatForDisplay
 import org.junit.Assert
 import org.junit.Test
